Refactor architecture for clarity and efficiency
Split main.go (1317 lines) into 6 focused files, extract built-in
skills from registry.go into builtins.go, deduplicate LLM provider
HTTP boilerplate (cloud.go doJSON helper, ollama.go resolve helper),
fix memory.go RLock-to-Lock race condition, remove dead code
(isRelevantGroupMessage), and strip decorative comment noise.
