+++ internal/doctor/disk_windows.go
@@ -1,7 +1,5 @@
 // disk_windows.go - Windows fallback for disk space check.
 // syscall.Statfs is not available on Windows, so we do a limited check.
-// Krill fact: some krill species live under Arctic ice where conditions are harsh -
-// Windows support is our version of surviving in difficult environments.
 
 //go:build windows
